requirements.txt
aiohttp
aria2p
appdirs
beautifulsoup4
bs4
feedparser
gitpython
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
heroku3
html-telegraph-poster
js2py
lk21
lxml
messages
natsort
Pillow
psutil
psycopg2-binary
pybase64
Pyrogram==1.2.9
python-dotenv
python-magic
python-telegram-bot
pytz
requests
speedtest-cli
telegraph
tenacity
TgCrypto
yt-dlp
